Dodajanje modela, prestrukturiranje
